# Replace image crate with faster jpg decoder and faster resize