Previous stats with 100 queries
Minimum request time: 437.515ms
Maximum request time: 773.3932ms
Average request time: 564.202971ms

Current stats with 100 queries
Minimum request time: 299.0822ms
Maximum request time: 518.6847ms
Average request time: 374.61788ms